1. remove index 2. reader include extract table
