+++ src/main/kotlin/wang/crick/ddl2plantuml/Table.kt
@@ -1,6 +1,5 @@
 package wang.crick.ddl2plantuml
 
-import wang.crick.ddl2plantuml.dictionary.IndexTypeEnum
 import java.io.Serializable
 
 /**
@@ -9,14 +8,10 @@ import java.io.Serializable
  * @author wangyuheng@outlook.com
  * @date 2019-01-20 12:54
  */
-class Table(var name: String, var comment: String, var columnList: List<Column>, var indexList: List<Index>) : Serializable
+class Table(var name: String, var comment: String, var columnList: List<Column>) : Serializable
 
 /**
  * 字段
  */
 data class Column(var name: String, var comment: String, var type: String, var size: Int?, var defaultValue: String?, var notNull: Boolean) : Serializable
 
-/**
- * 索引
- */
-class Index constructor(type: IndexTypeEnum, comment: String, columnList: List<Column?>)
