+++ src/main/kotlin/wang/crick/ddl2plantuml/Template.kt
@@ -1,4 +1,4 @@
-package wang.crick.ddl2plantuml.dictionary
+package wang.crick.ddl2plantuml
 
 /**
  * 模板
