--- src/main/kotlin/wang/crick/ddl2plantuml/dictionary/IndexTypeEnum.kt
@@ -1,12 +1,0 @@
-package wang.crick.ddl2plantuml.dictionary
-
-/**
- * 索引类型
- * @author wangyuheng
- * @date 2019-01-20 14:07
- */
-enum class IndexTypeEnum {
-    PRIMARY,
-    UNIQUE,
-    ORDINARY;
-}
